backend/python_automation_backend/requirements.txt
google-api-python-client==2.154.0
google-auth==2.36.0
requests==2.32.3
python-dotenv==1.0.1
